Move IoFolder implementation to header
